pyproject.toml
[project]
name = "agentic-baml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "baml-py>=0.204.0",
    "docling>=2.44.0",
    "fastapi>=0.116.1",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
    "websockets>=15.0.1",
]
